src/payments/requirements.txt
alembic==1.16.2
Audit-Alembic==0.2.0
kafka-python==2.2.15
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.1.1
PyYAML==6.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
